order-service/requirements.txt
django
djangorestframework
django-cors-headers
requests
psycopg2-binary
pika
